refactor: API 기능별 분리, 폴더 위치 이동
