+++ src/frontend/api/guestbook/type.ts
@@ -1,0 +1,13 @@
+export interface IAddGuestBookReqType {
+  content: string;
+}
+
+export interface IModifyGuestBookReqType {
+  content: string;
+}
+
+export interface IResType {
+  code: number;
+  message: string;
+  data: any;
+}
