+++ src/frontend/api/review/type.ts
@@ -1,0 +1,5 @@
+export interface IResType {
+  code: number;
+  message: string;
+  data: any;
+}
