+++ src/frontend/api/type.ts
@@ -1,0 +1,1 @@
+export interface IReqType {}
