+++ src/frontend/api/user/api.ts
@@ -1,0 +1,19 @@
+import axios from 'axios';
+import { IResType } from './type';
+import { axiosClient } from 'api/axiosClient';
+
+export async function getProfileAPI(userId: number) {
+  const response = await axiosClient.get<IResType>(
+    `${process.env.NEXT_PUBLIC_BASE_URL_MEMBERSHIP}/membership/profile/${userId}`,
+  );
+
+  return response.data;
+}
+
+export async function getWithFriendAPI(userId: number) {
+  const response = await axiosClient.get<IResType>(
+    `${process.env.NEXT_PUBLIC_BASE_URL_MEMBERSHIP}/membership/profile/${userId}/friends`,
+  );
+
+  return response.data;
+}
