+++ src/frontend/api/wishlist/type.ts
@@ -1,0 +1,13 @@
+export interface IResType {
+  code: number;
+  message: string;
+  data: any;
+}
+
+export interface IDoWishReqType {
+  game_id: number;
+}
+
+export interface IDoUnWishReqType {
+  game_id: number;
+}
