+++ src/frontend/modules/game/saga.ts
@@ -6,12 +6,10 @@ import {
   getCategoriesAPI,
   getGameAPI,
   getGamesByCategoryAPI,
-  doWishAPI,
   getUserDataAPI,
-  doUnWishAPI,
-  getWishListAPI,
   getGameInfoByIdListAPI,
-} from 'pages/api/game/api';
+} from 'api/game/api';
+import { doWishAPI, doUnWishAPI, getWishListAPI } from 'api/wishlist/api';
 import {
   getCategories,
   GET_CATEGORIES,
