+++ src/frontend/modules/game/sagaFunction.ts
@@ -1,4 +1,4 @@
-import { IAddCartInfoReqType, IRmCartInfoReqType } from 'pages/api/game/type';
+import { IAddCartInfoReqType, IRmCartInfoReqType } from 'api/game/type';
 
 export const addCartToStore = async ({ prev, game_id }: IAddCartInfoReqType) => {
   //중복제거 로직
@@ -10,6 +10,7 @@ export const addCartToStore = async ({ prev, game_id }: IAddCartInfoReqType) => 
     data: uniqueArr,
   };
 };
+
 export const rmCartToStore = async ({ prev, game_id }: IRmCartInfoReqType) => {
   //특정 index제거 로직
   const filtered = prev.filter((ele) => ele !== game_id);
