+++ src/frontend/modules/user/actions.ts
@@ -3,7 +3,7 @@ import { AxiosError } from 'axios';
 
 import { createRequestActionTypes } from 'modules/utils/actionUtils';
 
-import { IDoSignupReqType, IResType, IResStoreType, ISaveUserInfoReqtype } from 'pages/api/user/type';
+import { IDoSignupReqType, IResType, IResStoreType, ISaveUserInfoReqtype } from 'api/user/type';
 
 export const [DO_SIGNUP, DO_SIGNUP_SUCCESS, DO_SIGNUP_FAIL] = createRequestActionTypes('DO_SIGNUP');
 export const [SAVE_USERINFO, SAVE_USERINFO_SUCCESS, SAVE_USERINFO_FAIL] = createRequestActionTypes('SAVE_USERINFO');
