+++ src/frontend/modules/user/saga.ts
@@ -2,7 +2,7 @@ import { takeLatest } from 'redux-saga/effects';
 
 import createAsyncSaga from 'modules/utils/sagaUtils';
 
-import { doSignupAPI } from '../../pages/api/user/api';
+import { doSignupAPI } from 'api/auth/api';
 import { doSignup, DO_SIGNUP, saveUserInfo, SAVE_USERINFO } from 'modules/user/actions';
 import { saveUserInfoToStore } from 'modules/user/sagaFunction';
